Disambiguate CKAN.Utilities and Tests.Utilities
